pyproject.toml
[project]
name = "capstone-ingest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "country-converter>=1.3.2",
    "dbt-bigquery>=1.10.0",
    "dbt-core>=1.10.13",
    "geopandas>=1.1.3",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=2.4,<3.2",
    "idna==3.11",
    "matplotlib>=3.10.8",
    "numpy==2.4.2",
    "pandas==3.0.1",
    "psycopg[binary]>=3.2.9",
    "pyarrow==23.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "searoute>=1.4.3",
    "six==1.17.0",
    "urllib3==2.6.3",
]

[dependency-groups]
dev = [
    "dbt-bigquery>=1.10.0",
    "dbt-core>=1.11.7",
    "folium>=0.20.0",
]
